BE: fixed post carts/user/id route to have active cart
