+++ server/api/index.js
@@ -7,6 +7,8 @@ router.use('/plants', require('./plants'))
 
 router.use('/carts', require('./carts'))
 
+router.use('/checkout', require('./checkout'))
+
 router.use((req, res, next) => {
   const error = new Error('Not Found')
   error.status = 404
